separates out the message click functionality into a separate component and renames component to MessageListItem
